app/requirements.txt
Django==4.2.1
Pillow==9.5.0
django-import-export==3.2.0
openpyxl==3.1.2 
psycopg2==2.9.6
django-allauth==0.54.0
django-crispy-forms==2.0
xlwt==1.3.0
mailjet-rest==1.3.4
django-mailjet==0.3.1
crispy-bootstrap4==2022.1
gunicorn==20.1.0

